requirements.txt
Django>=5.0
djangorestframework>=3.15
django-filter>=24.2
Pillow>=10.0
djangorestframework-simplejwt>=5.3.0